Adding the bugfix, and new camera support for Axis 207M and 207MW from David Bird.  Thanks for your help David!  His email included with the patch follows:
Hi Jason,

Attached is a small patch for jipcam. I attempted to add support for
207M and 207MW - just for completeness (I'm using 207MW). But, more
importantly, there is a patch for a bug in CameraAPI.getImage(). 

Also, I noticed your blog entry for 'JipCam for sale?' ... As the
creator of jradius.net, I have given the topic some thought. Clearly,
there are various motives for open-source. For JRadius, it turned out
nice that work I did for a company that I can still use today for
consulting. It also served the Company well and was not directly part of
their business. 

It's great that JipCam is open-source. I believe it is highly beneficial
(especially for the hardware vendor) that device integration projects
like this are open-sourced and kept active. JipCam was the reason I
bought my Axis and am recommending it as the preferred camera to my
client. 

Cheers,
David